requirements.txt
django

djangorestframework
markdown
django-filter

celery[redis]

psycopg2-binary